chore: implemented new foldering structure
