+++ mythical-mysfit-api/src/api/models/mysfitModel.go
